Rename views package to view
