+++ src/main/java/website/fgto/royals_of_britain/view/Apology.java
@@ -1,11 +1,11 @@
-package website.fgto.royals_of_britain.views.houses.intros;
+package website.fgto.royals_of_britain.view;
 
 import org.apache.wicket.markup.html.panel.Panel;
 
-public class Stuart extends Panel {
+public class Apology extends Panel {
   private static final long serialVersionUID = 1L;
 
-  public Stuart(String id) {
+  public Apology(String id) {
     super(id);
   }
 }
