+++ src/main/java/website/fgto/royals_of_britain/view/Footer.html
