+++ src/main/java/website/fgto/royals_of_britain/view/Footer.java
@@ -1,4 +1,4 @@
-package website.fgto.royals_of_britain.views;
+package website.fgto.royals_of_britain.view;
 
 import org.apache.wicket.markup.html.basic.Label;
 import org.apache.wicket.markup.html.panel.Panel;
