+++ src/main/java/website/fgto/royals_of_britain/view/HomePage.java
@@ -1,4 +1,4 @@
-package website.fgto.royals_of_britain.views;
+package website.fgto.royals_of_britain.view;
 
 import org.apache.wicket.request.mapper.parameter.PageParameters;
 import org.apache.wicket.markup.html.basic.Label;
