+++ src/main/java/website/fgto/royals_of_britain/view/houses/intros/Bruce.java
@@ -1,11 +1,11 @@
-package website.fgto.royals_of_britain.views.houses.intros;
+package website.fgto.royals_of_britain.view.houses.intros;
 
 import org.apache.wicket.markup.html.panel.Panel;
 
-public class Wessex extends Panel {
+public class Bruce extends Panel {
   private static final long serialVersionUID = 1L;
 
-  public Wessex(String id) {
+  public Bruce(String id) {
     super(id);
   }
 }
