+++ src/main/java/website/fgto/royals_of_britain/view/houses/intros/Hanover.java
@@ -1,4 +1,4 @@
-package website.fgto.royals_of_britain.views.houses.intros;
+package website.fgto.royals_of_britain.view.houses.intros;
 
 import org.apache.wicket.markup.html.panel.Panel;
