+++ src/main/java/website/fgto/royals_of_britain/view/houses/intros/Tudor.java
@@ -1,11 +1,11 @@
-package website.fgto.royals_of_britain.views;
+package website.fgto.royals_of_britain.view.houses.intros;
 
 import org.apache.wicket.markup.html.panel.Panel;
 
-public class Apology extends Panel {
+public class Tudor extends Panel {
   private static final long serialVersionUID = 1L;
 
-  public Apology(String id) {
+  public Tudor(String id) {
     super(id);
   }
 }
