+++ src/main/java/website/fgto/royals_of_britain/view/houses/intros/Wessex.java
@@ -1,11 +1,11 @@
-package website.fgto.royals_of_britain.views.houses.intros;
+package website.fgto.royals_of_britain.view.houses.intros;
 
 import org.apache.wicket.markup.html.panel.Panel;
 
-public class Bruce extends Panel {
+public class Wessex extends Panel {
   private static final long serialVersionUID = 1L;
 
-  public Bruce(String id) {
+  public Wessex(String id) {
     super(id);
   }
 }
